requirements.txt
streamlit
pandas
numpy
opencv-python
scikit-learn
scikit-image
requests
Pillow